리포트 개인화
