+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.62.8",
     "@tosspayments/tosspayments-sdk": "^2.3.2",
     "axios": "^1.7.7",
     "body-parser": "^1.20.2",
